requirements.txt
flask
flask-socketio
nautilus_trader
pandas
numpy
xgboost
pyyaml
requests
numba
